requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.13.3
    # via gdown
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   osam
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   labelme
    #   loguru
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.16.1
    # via
    #   gdown
    #   torch
flatbuffers==25.2.10
    # via onnxruntime
fonttools==4.55.3
    # via matplotlib
fsspec==2024.12.0
    # via torch
gdown==5.2.0
    # via
    #   labelme
    #   osam
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via requests
imageio==2.37.0
    # via scikit-image
imgviz==1.7.6
    # via
    #   labelme
    #   osam
importlib-metadata==8.5.0
    # via build
jinja2==3.1.5
    # via torch
kiwisolver==1.4.8
    # via matplotlib
labelme==5.6.1
    # via -r requirements.in
lazy-loader==0.4
    # via scikit-image
loguru==0.7.3
    # via
    #   labelme
    #   osam
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.0
    # via
    #   -r requirements.in
    #   imgviz
    #   labelme
mpmath==1.3.0
    # via sympy
natsort==8.4.0
    # via labelme
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
numpy==2.2.1
    # via
    #   contourpy
    #   imageio
    #   imgviz
    #   labelme
    #   matplotlib
    #   onnxruntime
    #   opencv-python
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchvision
onnxruntime==1.20.1
    # via
    #   labelme
    #   osam
opencv-python==4.10.0.84
    # via -r requirements.in
osam==0.2.2
    # via labelme
packaging==24.2
    # via
    #   build
    #   lazy-loader
    #   matplotlib
    #   onnxruntime
    #   qtpy
    #   scikit-image
pillow==11.0.0
    # via
    #   imageio
    #   imgviz
    #   labelme
    #   matplotlib
    #   osam
    #   scikit-image
    #   torchvision
pip-tools==7.4.1
    # via -r requirements.in
protobuf==5.29.3
    # via onnxruntime
pydantic==2.10.6
    # via osam
pydantic-core==2.27.2
    # via pydantic
pyparsing==3.2.0
    # via matplotlib
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyqt5==5.15.11
    # via labelme
pyqt5-qt5==5.15.2
    # via pyqt5
pyqt5-sip==12.17.0
    # via pyqt5
pyreadline3==3.5.4
    # via humanfriendly
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.2
    # via
    #   imgviz
    #   labelme
qtpy==2.4.3
    # via labelme
requests[socks]==2.32.3
    # via gdown
scikit-image==0.25.1
    # via labelme
scipy==1.15.2
    # via scikit-image
six==1.17.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sympy==1.13.1
    # via
    #   onnxruntime
    #   torch
termcolor==2.5.0
    # via labelme
tifffile==2025.1.10
    # via scikit-image
tomli==2.2.1
    # via
    #   build
    #   pip-tools
torch[cu113]==2.6.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision[cu113]==0.21.0
    # via -r requirements.in
tqdm==4.67.1
    # via gdown
typing-extensions==4.12.2
    # via
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via requests
wheel==0.45.1
    # via pip-tools
win32-setctime==1.2.0
    # via loguru
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
